fix: remove theme from lib
