+++ src/components/Widgets/AnimatedIcons/Spinner.jsx
@@ -1,18 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import CircularProgress from '@material-ui/core/CircularProgress'
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles'
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#00A4FF'
-    },
-    secondary: {
-      main: '#ffffff'
-    }
-  }
-})
 
 export default class Spinner extends Component {
   static displayName = 'Spinner'
@@ -28,10 +16,6 @@ export default class Spinner extends Component {
   }
 
   render() {
-    return (
-      <MuiThemeProvider theme={theme}>
-        <CircularProgress {...this.props} />
-      </MuiThemeProvider>
-    )
+    return <CircularProgress {...this.props} />
   }
 }
