+++ src/components/Widgets/Button.jsx
@@ -1,19 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import MuiButton from '@material-ui/core/Button'
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles'
-
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      main: '#00A4FF',
-      contrastText: '#ffffff'
-    },
-    secondary: {
-      main: '#ffffff'
-    }
-  }
-})
 
 export default class Button extends Component {
   static displayName = 'Button'
@@ -38,15 +25,13 @@ export default class Button extends Component {
     const { children, secondary } = this.props
 
     return (
-      <MuiThemeProvider theme={theme}>
-        <MuiButton
-          {...this.props}
-          color={secondary ? 'secondary' : 'primary'}
-          variant="contained"
-        >
-          {children}
-        </MuiButton>
-      </MuiThemeProvider>
+      <MuiButton
+        {...this.props}
+        color={secondary ? 'secondary' : 'primary'}
+        variant="contained"
+      >
+        {children}
+      </MuiButton>
     )
   }
 }
